@@ -1,17 +1,14 @@
 {
   "name": "angular-shop",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "node server.js",
-    "startlocal": "ng serve",
+    "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "postinstall": "ng build --aot -prod",
-    "deploy": "git push && git push heroku"
+    "e2e": "ng e2e"
   },
   "engines": {
     "node": "8.11.1",
@@ -20,23 +17,21 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "^5.2.0",
-    "@angular/cli": "~1.7.3",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular-bootstrap-md": "^5.0.5",
+    "@angular/animations": "^6.0.3",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "angular-bootstrap-md": "^6.1.1",
     "angular2-toaster": "^5.0.1",
     "angularfire2": "5.0.0-rc.3",
     "animate.css": "^3.6.1",
     "bootstrap": "^4.1.1",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.4",
     "express": "^4.16.3",
     "firebase": "4.6.0",
     "font-awesome": "^4.7.0",
@@ -48,31 +43,32 @@
     "node-sass": "^4.9.0",
     "owl.carousel": "^2.3.4",
     "primeng": "^5.2.5",
-    "rxjs": "^5.5.6",
+    "rxjs-compat": "^6.1.0",
+    "rxjs": "^6.0.0",
     "shortid": "^2.2.8",
     "tether": "^1.4.4",
-
-    "zone.js": "^0.8.19"
+    "web-animations-js": "^2.3.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.6.6",
+    "@angular/cli": "~6.0.7",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "webpack": "^4.6.0"
+    "typescript": "~2.7.2"
   }
 }
