@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 import * as firebase from "firebase/app";
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 import { log } from "util";
 import { Buffer } from "buffer";
 import { User } from "../models/user";
