@@ -1,18 +1,13 @@
-// Core Dependencies
-import { RouterModule } from "@angular/router";
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule, NO_ERRORS_SCHEMA } from "@angular/core";
 
-// Configurations
-import { AppRoutes } from "./app.routing";
-
-// Components
 import { AppComponent } from "./app.component";
-import { LoaderSpinnerModule } from "./shared/loader-spinner/loader-spinner";
-import { SharedModule } from "./shared/shared.module";
 import { IndexModule } from "./index/index.module";
-import { UserModule } from "./user/user.module";
 import { ProductModule } from "./product/product.module";
+import { UserModule } from "./user/user.module";
+import { SharedModule } from "./shared/shared.module";
+import { RouterModule } from "@angular/router";
+import { AppRoutes } from "./app.routing";
 
 @NgModule({
   declarations: [AppComponent],
